+++ sdk/build.gradle
@@ -126,7 +126,6 @@ dependencies {
     testImplementation 'androidx.test.espresso:espresso-core:3.3.0-alpha01'
     testImplementation 'com.android.support:support-annotations:28.0.0'
     testImplementation "androidx.test.ext:junit:1.1.1"
-    testImplementation 'androidx.test.espresso:espresso-web:3.2.0'
 
     androidTestImplementation "org.robolectric:annotations:4.3"
     androidTestImplementation 'androidx.test:rules:1.3.0-alpha01'
@@ -134,7 +133,6 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0-alpha01'
     androidTestImplementation 'com.android.support:support-annotations:28.0.0'
     androidTestImplementation "androidx.test.ext:junit:1.1.1"
-    androidTestImplementation 'androidx.test.espresso:espresso-web:3.2.0'
 }
 
 repositories {
