+++ sdk/src/main/AndroidManifest.xml
@@ -2,4 +2,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="co.omise.android">
 
+    <uses-permission android:name="android.permission.INTERNET"/>
 </manifest>
